Avoiding race condition for immediate disconnect with data in flight
The socket event may have an operation pending completion, and should
not be returned to the socket event pool until it has fired. This happens
if a server indicates the connection should be disconnected immediately.

In unneeded IAsyncResult interface is removed. The component is no longer
used in that way.